refactor: renames "panel" package to "panels"
